Gauge: Drop v1 gauge
