+++ .gitignore
@@ -275,3 +275,4 @@ devenv/docker/blocks/auth/signer/keys/ec_private_key.pem
 .specs
 .playwright-mcp/
 overrides.yaml
+.gitnexus
