+++ eslint-suppressions.json
@@ -3814,16 +3814,6 @@
       "count": 1
     }
   },
-  "public/app/plugins/panel/gauge/v1/GaugeMigrations.ts": {
-    "@typescript-eslint/no-explicit-any": {
-      "count": 1
-    }
-  },
-  "public/app/plugins/panel/gauge/v1/GaugePanel.tsx": {
-    "react-prefer-function-component/react-prefer-function-component": {
-      "count": 1
-    }
-  },
   "public/app/plugins/panel/geomap/GeomapPanel.tsx": {
     "react-prefer-function-component/react-prefer-function-component": {
       "count": 1
