+++ public/app/plugins/panel/gauge/suggestions.test.ts
